+++ src/Symfony/Bundle/FrameworkBundle/Tests/DependencyInjection/ConfigurationTest.php
@@ -22,6 +22,7 @@ use Symfony\Component\Lock\Store\SemaphoreStore;
 use Symfony\Component\Mailer\Mailer;
 use Symfony\Component\Messenger\MessageBusInterface;
 use Symfony\Component\Notifier\Notifier;
+use Symfony\Component\RateLimiter\Policy\TokenBucketLimiter;
 use Symfony\Component\Uid\Factory\UuidFactory;
 
 class ConfigurationTest extends TestCase
@@ -443,6 +444,7 @@ class ConfigurationTest extends TestCase
                 'mapping' => ['paths' => []],
             ],
             'property_access' => [
+                'enabled' => true,
                 'magic_call' => false,
                 'magic_get' => true,
                 'magic_set' => true,
@@ -562,11 +564,11 @@ class ConfigurationTest extends TestCase
                 'private_headers' => [],
             ],
             'rate_limiter' => [
-                'enabled' => false,
+                'enabled' => !class_exists(FullStack::class) && class_exists(TokenBucketLimiter::class),
                 'limiters' => [],
             ],
             'uid' => [
-                'enabled' => class_exists(UuidFactory::class),
+                'enabled' => !class_exists(FullStack::class) && class_exists(UuidFactory::class),
                 'default_uuid_version' => 6,
                 'name_based_uuid_version' => 5,
                 'time_based_uuid_version' => 6,