+++ src/Symfony/Component/HttpKernel/EventListener/SessionListener.php
@@ -14,6 +14,7 @@ namespace Symfony\Component\HttpKernel\EventListener;
 use Psr\Container\ContainerInterface;
 use Symfony\Component\HttpFoundation\Session\SessionInterface;
 use Symfony\Component\HttpFoundation\Session\Storage\NativeSessionStorage;
+use Symfony\Component\HttpKernel\Event\RequestEvent;
 
 /**
  * Sets the session in the request.
@@ -33,10 +34,12 @@ class SessionListener extends AbstractSessionListener
         parent::__construct($container, $debug);
     }
 
-    protected function getSession(): ?SessionInterface
+    public function onKernelRequest(RequestEvent $event)
     {
-        if (!$this->container->has('session')) {
-            return null;
+        parent::onKernelRequest($event);
+
+        if (!$event->isMasterRequest() || !$this->container->has('session')) {
+            return;
         }
 
         if ($this->container->has('session_storage')
@@ -45,6 +48,13 @@ class SessionListener extends AbstractSessionListener
             && $masterRequest->isSecure()
         ) {
             $storage->setOptions(['cookie_secure' => true]);
+        }
+    }
+
+    protected function getSession(): ?SessionInterface
+    {
+        if (!$this->container->has('session')) {
+            return null;
         }
 
         return $this->container->get('session');